accelerate "953. Verifying an Alien Dictionary"
Using `String.indexOf()` is, in theory, less performant than doing a looking in a hash table, since `indexOf()` does a linear search on the `String`. However, it seems that for this specific dataset, the cost of building the hash table does not compensate for the faster look-up in the `String`.